requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
gspread
